refact routes file
