+++ server.js
@@ -1,17 +1,17 @@
 const express = require('express')
 const nunjucks = require('nunjucks')
+const routes = require('./routes')
 
 const methodOverride = require('method-override')
 
-const routes = require('./routes')
 
 
 const server = express()
 
 server.use(express.urlencoded({extended: true}))
 server.use(express.static('public'))
-server.use(routes)
 server.use(methodOverride('_method'))
+server.use(routes)
 
 
 server.set("view engine", "njk")
